+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/SearchOptions.kt
@@ -116,6 +116,14 @@ public class SearchOptions @JvmOverloads public constructor(
      * Threshold specified in meters.
      */
     public val indexableRecordsDistanceThresholdMeters: Double? = null,
+
+    /**
+     * Besides the basic metadata attributes, developers can request additional
+     * attributes by setting attribute_sets parameter with attribute set values,
+     * for example &attribute_sets=basic,photos,visit.
+     * The requested metadata will be provided in metadata object in the response.
+     */
+    public val attributeSets: List<AttributeSet>? = null,
 ) : Parcelable {
 
     init {
@@ -144,6 +152,7 @@ public class SearchOptions @JvmOverloads public constructor(
         unsafeParameters: Map<String, String>? = this.unsafeParameters,
         ignoreIndexableRecords: Boolean = this.ignoreIndexableRecords,
         indexableRecordsDistanceThresholdMeters: Double? = this.indexableRecordsDistanceThresholdMeters,
+        attributeSets: List<AttributeSet>? = this.attributeSets,
     ): SearchOptions {
         return SearchOptions(
             proximity = proximity,
@@ -160,6 +169,7 @@ public class SearchOptions @JvmOverloads public constructor(
             unsafeParameters = unsafeParameters,
             ignoreIndexableRecords = ignoreIndexableRecords,
             indexableRecordsDistanceThresholdMeters = indexableRecordsDistanceThresholdMeters,
+            attributeSets = attributeSets,
         )
     }
 
@@ -193,6 +203,7 @@ public class SearchOptions @JvmOverloads public constructor(
         if (unsafeParameters != other.unsafeParameters) return false
         if (ignoreIndexableRecords != other.ignoreIndexableRecords) return false
         if (!indexableRecordsDistanceThresholdMeters.safeCompareTo(other.indexableRecordsDistanceThresholdMeters)) return false
+        if (attributeSets != other.attributeSets) return false
 
         return true
     }
@@ -215,6 +226,7 @@ public class SearchOptions @JvmOverloads public constructor(
         result = 31 * result + (unsafeParameters?.hashCode() ?: 0)
         result = 31 * result + ignoreIndexableRecords.hashCode()
         result = 31 * result + indexableRecordsDistanceThresholdMeters.hashCode()
+        result = 31 * result + (attributeSets?.hashCode() ?: 0)
         return result
     }
 
@@ -237,6 +249,7 @@ public class SearchOptions @JvmOverloads public constructor(
                 "unsafeParameters=$unsafeParameters, " +
                 "ignoreIndexableRecords=$ignoreIndexableRecords, " +
                 "indexableRecordsDistanceThresholdMeters=$indexableRecordsDistanceThresholdMeters" +
+                "attributeSets=$attributeSets" +
                 ")"
     }
 
@@ -260,6 +273,7 @@ public class SearchOptions @JvmOverloads public constructor(
         private var unsafeParameters: Map<String, String>? = null
         private var ignoreIndexableRecords: Boolean = false
         private var indexableRecordsDistanceThresholdMeters: Double? = null
+        private var attributeSet: List<AttributeSet>? = null
 
         internal constructor(options: SearchOptions) : this() {
             proximity = options.proximity
@@ -276,6 +290,7 @@ public class SearchOptions @JvmOverloads public constructor(
             unsafeParameters = options.unsafeParameters
             ignoreIndexableRecords = options.ignoreIndexableRecords
             indexableRecordsDistanceThresholdMeters = options.indexableRecordsDistanceThresholdMeters
+            attributeSet = options.attributeSets
         }
 
         /**
@@ -395,6 +410,16 @@ public class SearchOptions @JvmOverloads public constructor(
         }
 
         /**
+         * Besides the basic metadata attributes, developers can request additional
+         * attributes by setting attribute_sets parameter with attribute set values,
+         * for example &attribute_sets=basic,photos,visit.
+         * The requested metadata will be provided in metadata object in the response.
+         */
+        public fun attributeSet(attributeSet: List<AttributeSet>?): Builder = apply {
+            this.attributeSet = attributeSet
+        }
+
+        /**
          * Create [SearchOptions] instance from builder data.
          */
         public fun build(): SearchOptions = SearchOptions(
@@ -412,6 +437,7 @@ public class SearchOptions @JvmOverloads public constructor(
             unsafeParameters = unsafeParameters,
             ignoreIndexableRecords = ignoreIndexableRecords,
             indexableRecordsDistanceThresholdMeters = indexableRecordsDistanceThresholdMeters,
+            attributeSets = attributeSet
         )
     }
 }
@@ -449,6 +475,7 @@ internal fun SearchOptions.mapToCore(): CoreSearchOptions = CoreSearchOptions(
     routeOptions?.deviation?.sarType?.rawName,
     routeOptions?.timeDeviationMinutes,
     unsafeParameters?.let { (it as? HashMap) ?: HashMap(it) },
+    attributeSets?.map { it.mapToCore() }
 )
 
 @JvmSynthetic
@@ -480,4 +507,5 @@ internal fun CoreSearchOptions.mapToPlatform(): SearchOptions = SearchOptions(
     unsafeParameters = addonAPI,
     ignoreIndexableRecords = ignoreUR,
     indexableRecordsDistanceThresholdMeters = urDistanceThreshold,
+    attributeSets = attributeSets?.map { it.mapToPlatform() },
 )