+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog for the Mapbox Search SDK for Android
 
+## [Unreleased]
+- Adding support for `AttributeSet`s when performing a retrieve
+- Adding additional fields to `SearchResultMetaData`
+
+### Mapbox dependencies
+- Search Native SDK `2.3.0-rc.4`
+- Common SDK `v24.6.0-rc1`
+
 ## 2.3.1
 
 ### Mapbox dependencies